Fix type declaration import issue
