+++ src/assets/theme/index.ts
@@ -1,7 +1,26 @@
 import { useContext } from "react";
-import { ThemeContext } from "react-native-elements";
+import { FullTheme, ButtonProps, ThemeContext } from "react-native-elements";
 import { ThemeProps } from "react-native-elements/dist/config/ThemeProvider";
-import { Theme } from "../../@types/theme";
+import { Colors } from "react-native-elements/dist/config/colors";
+import { RecursivePartial } from "utils/core";
+
+interface IExtendedThemeButtonProps extends Partial<ButtonProps> {
+    titleAppearances?: {
+        titleColors?: {
+            primary: string;
+        };
+        fontFamily?: string;
+    };
+}
+
+interface IExtendedThemeColorProps extends RecursivePartial<Colors> {
+    readonly darkBg: string;
+}
+
+export interface Theme extends Partial<FullTheme> {
+    Button?: IExtendedThemeButtonProps;
+    colors?: IExtendedThemeColorProps;
+}
 
 export const useThemeContext = () => useContext<ThemeProps<Theme>>(ThemeContext);
 
@@ -19,7 +38,5 @@ const theme: Theme = {
         darkBg: "#0F0F0F",
     },
 };
-
-export * from "../../@types/theme";
 
 export default theme;
